requirements.txt
streamlit
chromadb
langchain
ollama
python-telegram-bot
